custom config errors
